Moved the resolution and other settings to a new script
