+++ Assets/Prefabs/Settings.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 6c368f8c8f8004b48a36f14b1b397b5c
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
